whew

+++ src/org/mozilla/types/Int32Array.java
@@ -1,11 +1,9 @@
 package org.mozilla.types;
 
-import net.three4j.unit.utils.TypedArray;
-
 public class Int32Array extends TypedArray {
 
 	public static int BYTES_PER_ELEMENT = 4;
-	public static final String name = "Int16Array";
+	public static final String name = "Int32Array";
 
 	public Int32Array(double[] array) {
 		super(array.length);
